refactor: 건물 위키 파일 이동
각 building 컴포넌트에서 wiki 서브 디렉토리 아래로 이동

Related to: #117
