+++ src/pages/building/wiki/EditBuildingWiki.jsx
@@ -1,10 +1,10 @@
 import { useEffect, useRef, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
-import axios_api from "../../lib/axios_api";
+import axios_api from "../../../lib/axios_api";
 import { BUILDING_WIKI_BASE_PATH } from "./GetBuildingWiki";
 import $ from "jquery";
-import Footer from "../../components/common/Footer";
-import styles from "../../assets/css/module/building/EditBuildingWiki.module.css";
+import Footer from "../../../components/common/Footer";
+import styles from "../../../assets/css/module/building/wiki/EditBuildingWiki.module.css";
 import { Button } from "react-bootstrap";
 
 export default function EditBuildingWiki() {
