+++ src/routes/AppRoutes.jsx
@@ -68,8 +68,8 @@ import FeedChartPage from '../pages/feed/FeedChartPage';
 import FeedBuildingListPage from '../pages/feed/FeedBuildingListPage';
 import FeedListHomePage from '../pages/feed/FeedListHomePage';
 import GetMemberProfile from "../pages/member/GetMemberProfile";
-import GetBuildingWiki from "../pages/building/GetBuildingWiki";
-import EditBuildingWiki from "../pages/building/EditBuildingWiki";
+import GetBuildingWiki from "../pages/building/wiki/GetBuildingWiki";
+import EditBuildingWiki from "../pages/building/wiki/EditBuildingWiki";
 import ApplicantSample from "../pages/building/components/ApplicantSample";
 import FeedVoteForm from "../pages/feed/FeedVoteFormPage";
